Move common.fetcher_models to common.configs.fetcher
