+++ hasher-matcher-actioner/hmalib/common/tests/test_writebacker.py
@@ -7,7 +7,7 @@ import os
 from hmalib.lambdas.actions.writebacker import lambda_handler
 from hmalib.common.classification_models import WritebackTypes
 from hmalib.common.message_models import MatchMessage, WritebackMessage, BankedSignal
-from hmalib.common.fetcher_models import ThreatExchangeConfig
+from hmalib.common.configs.fetcher import ThreatExchangeConfig
 from hmalib.common import config as hmaconfig
