+++ hasher-matcher-actioner/hmalib/common/threatexchange_config.py
@@ -8,8 +8,7 @@ from hmalib.aws_secrets import AWSSecrets
 from threatexchange.api import ThreatExchangeAPI
 from hmalib.common.logging import get_logger
 from hmalib.common import config as hmaconfig
-from hmalib.common.config import HMAConfig
-from hmalib.common.fetcher_models import ThreatExchangeConfig
+from hmalib.common.configs.fetcher import ThreatExchangeConfig
 from botocore.exceptions import ClientError
 
 logger = get_logger(__name__)
